+++ src/app/voting-documents/wizard/components/manual-voting-card-voter-edit/manual-voting-card-voter-edit.component.ts
@@ -25,8 +25,8 @@ export class ManualVotingCardVoterEditComponent {
   public readonly salutations: EnumItemDescription<Salutation>[] = [];
   public readonly religions: EnumItemDescription<Religion>[] = [];
 
+  private configurationValue!: VotingCardLayoutDataConfiguration;
   public voterValue!: ManualVotingCardVoter;
-  public options!: VotingCardLayoutDataConfiguration;
   public swiss = true;
   public dateOfBirth?: string;
   public printEmpty = false;
@@ -43,9 +43,12 @@ export class ManualVotingCardVoterEditComponent {
 
   @Input()
   public set configuration(c: VotingCardLayoutDataConfiguration) {
-    this.options = c;
-    this.dateOfBirthIsValid = !c.includeDateOfBirth;
-    this.religionIsValid = !c.includeReligion;
+    this.configurationValue = c;
+    this.updateIsValidFields();
+  }
+
+  public get configuration(): VotingCardLayoutDataConfiguration {
+    return this.configurationValue;
   }
 
   @Input()
@@ -53,6 +56,7 @@ export class ManualVotingCardVoterEditComponent {
     this.voterValue = v;
     this.swiss = v.country.iso2 === 'CH';
     this.dateOfBirth = toBcDate(v.dateOfBirth);
+    this.updateIsValidFields();
   }
 
   constructor() {
@@ -67,12 +71,15 @@ export class ManualVotingCardVoterEditComponent {
 
   public updateSwissZipCode(e?: number): void {
     this.voterValue.swissZipCode = e;
-    // Can be removed if jira BC-766 is fixed.
-    this.swissZipCodeIsValid = !!this.voterValue.swissZipCode && this.voterValue.swissZipCode >= 1000;
+    this.updateIsValidFields();
+  }
+
+  public updateForeignZipCode(e?: string): void {
+    this.voterValue.foreignZipCode = e;
+    this.updateIsValidFields();
   }
 
   public updateDateOfBirth(birthdate?: string): void {
-    this.dateOfBirthIsValid = !!birthdate || this.configuration?.includeDateOfBirth;
     if (!birthdate) {
       delete this.voterValue.dateOfBirth;
       delete this.dateOfBirth;
@@ -80,11 +87,13 @@ export class ManualVotingCardVoterEditComponent {
       this.dateOfBirth = birthdate;
       this.voterValue.dateOfBirth = fromBcDate(this.dateOfBirth);
     }
+
+    this.updateIsValidFields();
   }
 
   public updateReligionCode(religion: Religion): void {
-    this.religionIsValid = !!religion || this.configuration?.includeReligion;
     this.voterValue.religion = religion ?? Religion.RELIGION_UNSPECIFIED;
+    this.updateIsValidFields();
   }
 
   public updateSwiss(b: boolean): void {
@@ -97,5 +106,22 @@ export class ManualVotingCardVoterEditComponent {
       this.voterValue.foreignZipCode = '';
       this.voterValue.country = { iso2: '', name: '' };
     }
+
+    this.updateIsValidFields();
+  }
+
+  private updateIsValidFields(): void {
+    // Can be removed if jira BC-766 is fixed.
+    this.swissZipCodeIsValid =
+      !this.swiss || (!!this.voterValue.swissZipCode && this.voterValue.swissZipCode >= 1000 && this.voterValue.swissZipCode <= 9999);
+
+    if (!this.configuration) {
+      this.religionIsValid = false;
+      this.dateOfBirthIsValid = false;
+      return;
+    }
+
+    this.religionIsValid = !!this.voterValue.religion || !this.configuration.includeReligion;
+    this.dateOfBirthIsValid = !!this.voterValue.dateOfBirth || !this.configuration.includeDateOfBirth;
   }
 }